Fixed compile warnings against Linux 2.6.12.1.
git-svn-id: https://svn.code.sf.net/p/linux-gpib/code/trunk@1279 b61deed1-901b-0410-a6dd-e94ecbdaea89

+++ linux-gpib/driver/ines/ines.h
@@ -92,11 +92,11 @@ void ines_set_xfer_counter( ines_private_t *priv, unsigned int count );
 /* inb/outb wrappers */
 static inline unsigned int ines_inb( ines_private_t *priv, unsigned int register_number )
 {
-	return inb( priv->nec7210_priv.iobase + register_number * priv->nec7210_priv.offset );
+	return inb((unsigned long)(priv->nec7210_priv.iobase) + register_number * priv->nec7210_priv.offset );
 }
 static inline void ines_outb( ines_private_t *priv, unsigned int value, unsigned int register_number )
 {
-	outb( value, priv->nec7210_priv.iobase + register_number * priv->nec7210_priv.offset );
+	outb( value, (unsigned long)(priv->nec7210_priv.iobase) + register_number * priv->nec7210_priv.offset );
 }
 
 // pcmcia init/cleanup
